feat: add debug logging
This patch adds debug logging to PineFlash, to aid with debugging and
creating bug reports.
This commit is dependant on aaa771b5, as it includes logging specific to
that patch.

This patch adds `env_logger` as a dependency; however, `log` was already
a transitive dependency, and this patch will expose the logging
functionality of dependencies automatically.

This patch adds some initial debug and trace logging statements to
PineFlash itself (specifically building on aaa771b5), and opens the door
for future contributions to also include useful diagnostic and
informational log messages.

It also exposes the logging functionality of dependencies.
INFO (or higher) messages from dependencies will be now be printed
automatically, and running PineFlash with the environment variable
`RUST_LOG=debug` or `RUST_LOG=trace` set will already print many
potentially useful diagnostic messages from dependencies.

The goal of this patch is to introduce new tools to aid in debugging,
and allow contributors to polish hasty println!() debugging into
reusable diagnostic messages where useful, rather than simply deleting
it to be recreated when needed again.

Related-to: gh-92