feat: ensure at least one stable version is offered
Previously, PineFlash would fetch the three latest version of IronOS
from GitHub, irrespective of the prerelease status of those versions.

This meant that PineFlash would only offer unstable versions if there
were three or more prereleases for the upcoming version.

As of 2025-03-03, PineFlash only offers:
- v2.23-rc3
- v2.23-rc2
- v2.23-rc1

This patch implements the following rules for choosing which versions of
IronOS to offer to the user:
- Releases marked as 'draft' on GitHub must never be offered
- At least one stable version must always be offered
- Prereleases of versions that have since been fully released should not
  be offered

With this patch, PineFlash instead offers the following versions on 2025-03-03:
- v2.23-rc3
- v2.23-rc2
- v2.22

Resolves: gh-92